+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Logging spans with unique id for priority plugin and slurm epilog collector (helps differentiate different runs in logs) ([@stefan-k](https://github.com/stefan-k)).
+
 ## [0.0.2] - 2022-08-01
 ### Added
 - Documentation of priority plugin on website ([@stefan-k](https://github.com/stefan-k)).